WW-5352 Repackage ParametersInterceptor and related classes
