+++ core/src/main/java/com/opensymphony/xwork2/interceptor/ParameterNameAware.java
@@ -19,12 +19,10 @@
 package com.opensymphony.xwork2.interceptor;
 
 /**
- * This interface is implemented by actions that want to declare acceptable parameters. Works in conjunction with {@link
- * ParametersInterceptor}. For example, actions may want to create a white list of parameters they will accept or a
- * blacklist of parameters they will reject to prevent clients from setting other unexpected (and possibly dangerous)
- * parameters.
+ * @deprecated since 6.4.0, use {@link org.apache.struts2.interceptor.parameter.ParameterNameAware}.
  */
-public interface ParameterNameAware {
+@Deprecated
+public interface ParameterNameAware extends org.apache.struts2.interceptor.parameter.ParameterNameAware {
 
     /**
      * Tests if the the action will accept the parameter with the given name.
@@ -33,5 +31,5 @@ public interface ParameterNameAware {
      * @return <tt>true</tt> if accepted, <tt>false</tt> otherwise
      */
     boolean acceptableParameterName(String parameterName);
-    
+
 }
